Reorganize GitHub handlers for long-term wiki support
This is a highly desirable feature, but the API doesn't quite support it yet. Disappointing
